Fix DoubleDouble equality check
When the doubledouble was low ( before 1E-12) the equality test was always returning true.
Moreover the equals was not checking the sum of hi+lo.

Fixing this bug raised some issues in term of precision in the existing tests.
Acceptance values in the tests were defined for standard double: that mean loss of accuracy.
Fixing this lack of accuracy was shown a loss of accuracy in the complex log calculation.

For some test the apfloat library was been introduced to compute the expected complex values with arbitraty precision (higher than the 30 digits precision of DoubleDouble)